requirements.txt
appdirs==1.4.4
asn1crypto==1.4.0
CacheControl==0.12.6
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==3.3.2
distlib==0.3.3
distro==1.6.0
enum34==1.1.10
future==0.18.2
html5lib==1.1
idna==2.10
importlib-metadata==2.1.1
ipaddress==1.0.23
more-itertools==5.0.0
msgpack==0.6.2
numpy==1.16.6
packaging==20.9
pathlib2==2.3.7.post1
pefile==2019.4.18
pep517==0.11.0
ply==3.11
progress==1.6
pyasn1==0.4.8
pycairo==1.18.2
pycparser==2.21
pycrypto==2.6.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytoml==0.1.21
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
scandir==1.10.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
webencodings==0.5.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
zenmap==7.92
zipp==1.1.1
